# openapi.yaml
basePath: /
consumes:
- application/json
definitions:
...
  Error:
    properties:
      error:
        description: Felmeddelande
        type: string
    type: object
...
  Object:
    properties:
      created_at:
        format: date-time
        type: string
      data:
        description: Fältvärden (nyckel = fältnamn)
        type: object
      id:
        type: integer
      id_full:
        description: Unikt sammansatt ID (t.ex. PROD-1.v1)
        type: string
      main_id:
        type: string
      object_type:
        type: string
      object_type_id:
        type: integer
      status:
        type: string
      updated_at:
        format: date-time
        type: string
      version:
        type: string
    type: object
...
info:
  contact:
    name: Ondfrukt
    url: https://github.com/ondfrukt/byggdelsuppdelning---demo
  description: REST API för Byggdelsuppdelning – ett PLM-system (Product Lifecycle
    Management) för hantering av byggdelar, objekttyper, relationer, dokument, klassificering
    och ändringshantering. Alla endpoints börjar med `/api/`.
  title: Byggdelsuppdelning API
  version: 1.0.0
paths:
...
  /api/search:
    get:
      parameters:
      - description: Sökterm (söker i ID och textfältvärden)
        in: query
        name: q
        required: true
        type: string
      - description: Filtrera på objekttypens namn
        in: query
        name: type
        required: false
        type: string
      - description: Begränsa sökning till specifikt fältnamn
        in: query
        name: field
        required: false
        type: string
      responses:
        '200':
          description: Matchande objekt
          schema:
            items:
              $ref: '#/definitions/Object'
            type: array
        '500':
          description: Sökningen misslyckades
          schema:
            $ref: '#/definitions/Error'
      summary: Search across all objects
      tags:
      - Search & Stats
  /api/stats:
    get:
      responses:
        '200':
          description: Statistik över objekt per typ
          schema:
            properties:
              objects_by_type:
                additionalProperties:
                  type: integer
                description: Antal objekt per objekttyp (nyckel = typnamn)
                type: object
              recent_objects:
                description: De 10 senast skapade objekten
                items:
                  $ref: '#/definitions/Object'
                type: array
              total_objects:
                description: Totalt antal objekt
                type: integer
            type: object
        '500':
          description: Serverfel
          schema:
            $ref: '#/definitions/Error'
      summary: Get statistics about the object-based system
      tags:
      - Search & Stats
...
produces:
- application/json
schemes:
- https
- http
securityDefinitions: {}
swagger: '2.0'
...